refactor : 클래스 명 변경 및 도메인에서 예외 진행
