+++ src/test/java/domain/NameTest.java
@@ -13,12 +13,12 @@ public class NameTest {
     @ParameterizedTest
     @ValueSource(strings = {"pobi", "crong", "jay", "jamie", "odo"})
     @DisplayName("생성자 이름 예외 테스트 : 성공 경우)")
-    void test_name_constructor_success(String carName) {
+    void test_name_constructor_success(String input) {
         // when
-        Car car = new Car(carName, 0);
+        CarName carName = new CarName(input);
 
         // then
-        assertThat(car.getCarName()).isEqualTo(carName);
+        assertThat(carName.getName()).isEqualTo(carName.getName());
     }
 
     @ParameterizedTest
@@ -29,20 +29,20 @@ public class NameTest {
         String expectedMessage = "자동차 이름은 5자 이하로 작성해주세요.";
 
         // when & then
-        assertThatThrownBy(() -> new Car(carName, 0))
+        assertThatThrownBy(() -> new CarName(carName))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessageContaining(expectedMessage);
     }
 
     @Test
-    @DisplayName("생성자 이름 예외 테스트 : 실패 경우 (이름이 공백)")
+    @DisplayName("생성자 이름 예외 테스트 : 실패 경우 (이름이 null)")
     void test_name_constructor_fail_empty_name() {
         // given
-        String carName = "";
+        String carName = null;
         String expectedMessage = "자동차를 한 대 이상 작성해주세요.";
 
         // when & then
-        assertThatThrownBy(() -> new Car(carName, 0))
+        assertThatThrownBy(() -> new CarName(carName))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessageContaining(expectedMessage);
     }
